Use recursive mutex in ccConsole.
This quick hack fixes a problem on my laptop where CloudCompare's
console logger was trying to lock the same mutex twice. Happily this was
occurring in the same thread, allowing use of a recursive mutex to lock
the mutex multiple times. It must be unlocked the same number of times
before another thread can lock it.